requirements.txt
streamlit
pymongo
pymupdf
python-docx
openai
bcrypt
fuzzywuzzy